+++ demo/components/popover/demo.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+  templateUrl: 'demo.component.html',
+})
+export class PopoverDemoComponent {
+  style = {
+    border: '3px double red',
+    padding: '20px',
+    overflow: 'hidden'
+  };
+}