+++ src/ng-vcl/popover/index.ts
@@ -3,6 +3,8 @@ import { CommonModule } from '@angular/common';
 import { PopoverComponent } from './popover.component';
 import { VCLOffClickModule } from '../off-click/index';
 
+export { PopoverComponent }
+
 @NgModule({
   imports: [
     CommonModule,
@@ -14,6 +16,3 @@ import { VCLOffClickModule } from '../off-click/index';
 })
 
 export class VCLPopoverModule { }
-export { PopoverComponent }
-
-